+++ implementations/slims/slims/protocol/mt/messagetype.go
