+++ implementations/slims/slims/vk/errors.go
